requirements.txt
aiohttp==3.13.3
demucs==4.0.1
docopt==0.6.2
hydra-core==1.3.2
ipython==8.12.3
nbconvert==7.17.0
pip==26.0.1
polars==1.38.1
pytorch-lightning==2.6.1
setuptools==79.0.1
tinycss2==1.4.0
torchcodec==0.10.0
yarg==0.1.9
